requirements.txt
fastapi
uvicorn
pydantic
httpx
Pillow
cloudinary
python-dotenv
resend
openai
